pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "characterization_utilities"
description = "Plugin to treat characterization steps and standardize them from proprietary data formats to NeXuS"
version = "1.4.0"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Matteo Bontorno", email = "mbontorno@fbk.eu" },
]
maintainers = [
    { name = "Matteo Bontorno", email = "mbontorno@fbk.eu" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
    "fabrication-utilities @ git+https://github.com/Trog-404/Fabrication-utilities.git",
    "pynxtools",
]

[project.urls]
Repository = "https://github.com/Trog-404/Characterization-utilities.git"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLC0415", # `import` should be at the top-level of a file
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']

# Schema packages entry points

characterization_entry_point = "characterization_utilities.schema_packages:Characterization_entry_point"
dataconverter_entry_point = "characterization_utilities.schema_packages:Dataconverter_entry_point"

## Specific technique entry points

em_schema_package_entry_point = "characterization_utilities.schema_packages:Em_schema_package_entry_point"

#app_entry_point = "characterization_utilities.apps:app_entry_point"

# Example uploads entry points

example_upload_entry_point = "characterization_utilities.example_uploads:example_upload_entry_point"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
